Tenatively working make_climo script.
I think I have a working make_climo script now, that should be
capable of calculating climatological seasonal means, and also of
calculating the standard deviation from those means. This script
calculates a mean for a given season for each year of input, and
then calculates the mean climatology and standard deviation from
the time series of seasonal averages. This is what makes the most
sense to me when I think of a "typical DJF", for example. In order
to accomplish this I had to modify the get_datetimes function that
manually decodes times in cesmutils. This function now decodes the
times using the time (and preferably the time_bnds) fields, taking
the base time from the units string. This has not been extensively
tested, but it seems to work so far.